Finished commenting

+++ Hydrogen/Window.cpp
@@ -42,7 +42,7 @@ void Window::beginRendering() {
 
 void Window::endRendering() {
 	Window::renderSpr.setTexture(this->rTex->getTexture());
-	Window::renderSpr.setPosition(sf::Vector2f(0.f, this->getSize().y));
+	Window::renderSpr.setPosition(sf::Vector2f(0.f, (float)this->getSize().y));
 	Window::renderSpr.setScale(1.f, -1.f);
 	this->window->draw(Window::renderSpr);
 	this->window->display();
